anchorhandle now in graph2
